+++ jkorm/src/main/kotlin/com/tang/jkorm/session/factory/SqlSessionFactoryBuilder.kt
@@ -6,6 +6,8 @@ import com.tang.jkorm.io.Resources
 import com.tang.jkorm.session.Configuration
 import com.tang.jkorm.session.factory.defaults.DefaultSqlSessionFactory
 import com.tang.jkorm.sql.defaults.DefaultSqlProviderFactory
+import com.tang.jkorm.transaction.TransactionFactory
+import com.tang.jkorm.transaction.jdbc.JdbcTransactionFactory
 import java.io.InputStream
 import javax.sql.DataSource
 
@@ -13,6 +15,8 @@ import javax.sql.DataSource
  * @author Tang
  */
 class SqlSessionFactoryBuilder {
+
+    lateinit var transactionFactory: TransactionFactory
 
     fun build(inputStream: InputStream): SqlSessionFactory {
         val datasource = Resources.getDataSourceProperties(inputStream)
@@ -23,8 +27,11 @@ class SqlSessionFactoryBuilder {
     fun build(dataSource: DataSource): SqlSessionFactory {
         val url = dataSource.connection.use { it.metaData.url }
         val sqlProvider = DefaultSqlProviderFactory().newSqlProvider(url)
+        if (!::transactionFactory.isInitialized) {
+            transactionFactory = JdbcTransactionFactory()
+        }
         printBanner()
-        return DefaultSqlSessionFactory(Configuration(dataSource, sqlProvider))
+        return DefaultSqlSessionFactory(Configuration(dataSource, sqlProvider, transactionFactory))
     }
 
     fun printBanner() {